MelodicFunctions

Documentation

data Progression

data Mod

Given duration, a, mode and a list of chords (in scale given in a + mode), returns progression. Example: progress hn (Bf,oo 4) Maj [I,IV,V,I] Note: Change tempo by using "tempo r m" (where r - ratio x/60bpm, m - melody).

oo :: Int -> Octave

Function wrapping Int into Octave.

heynow :: Music Pitch

accMaj :: Dur -> a -> Music a

Given duration and a (PitchClass,Octave) tuple, returns a Maj chord based on given pitch. Example: accMaj (5/8) (F,oo 2)

accMin :: Dur -> a -> Music a

Given duration and a, returns a Min chord based on a given pitch. Example: accMin sn (Ds,oo 4)

accDim :: Dur -> a -> Music a

Given duration and a, returns a diminished chord based on a given pitch. Example: accDim qn (Fss,oo 3)

composeMelody :: a -> Int -> Music a

Function which creates pseudo-random melody, based on given seed and key. Combines 8 tacts (2/4 measure) of melody, matching (kind of) the progression.

composeWhole :: a -> Int -> Music a

Function which combines random melody with a simple progression (lasting 8 tacts)

run2 :: Music Pitch